Create a conda env spec file with pkg versions stripped.
Build the image with the *-no-version* pkg spec to pull in the
latest version of each package along with new dependencies. Run
the resulting image and export the environment. Overwrite
conda-env.yml with the new environment file, commit the change, and
build the image again.

NB: The CPython package does have its version specified in this file.