feature: Added support to run Jupyter Notebook/Lab from the image.
Most of the shell scripts and Dockerfile changes are borrowed from
the Jupyter project Docker stacks with minor modifications. The
trouble is the jupyter server processes are throwing 403 Forbidden
errors when the user attempts to stop them. This is a minor
inconvenience, but it should not happen.

Added explanations and sample commands desribing how to start a
container from the image using Jupyter Lab/Notebook.